neuroassist/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
groq==0.9.0
httpx==0.27.2
langchain-nvidia-ai-endpoints
sentence-transformers==2.7.0
numpy==1.26.4
scipy==1.13.1
nibabel==5.2.1
pydicom==2.4.4
matplotlib==3.8.4
SimpleITK==2.3.1
scikit-learn==1.5.1
tqdm==4.66.5
tensorflow==2.16.1
